PART_01/requirements.txt
openai==1.52.0
python-dotenv==1.0.1